+++ packages/config/src/configs/moonbeam.ts
@@ -1,6 +1,7 @@
 import { BalanceBuilder, ContractBuilder } from '@moonbeam-network/xcm-builder';
 import {
   aca,
+  apillon,
   aseed,
   astr,
   bnc,
@@ -26,6 +27,7 @@ import {
   pha,
   pink,
   ring,
+  stink,
   sub,
   usdc,
   usdcwh,
@@ -506,6 +508,36 @@ export const moonbeamConfig = new ChainConfig({
     }),
     new AssetConfig({
       asset: ded,
+      balance: BalanceBuilder().substrate().assets().account(),
+      contract: ContractBuilder().Xtokens().transferMultiCurrencies(),
+      destination: polkadotAssetHub,
+      destinationFee: {
+        amount: 0.2,
+        asset: usdt,
+        balance: BalanceBuilder().substrate().assets().account(),
+      },
+      fee: {
+        asset: glmr,
+        balance: BalanceBuilder().substrate().system().account(),
+      },
+    }),
+    new AssetConfig({
+      asset: stink,
+      balance: BalanceBuilder().substrate().assets().account(),
+      contract: ContractBuilder().Xtokens().transferMultiCurrencies(),
+      destination: polkadotAssetHub,
+      destinationFee: {
+        amount: 0.2,
+        asset: usdt,
+        balance: BalanceBuilder().substrate().assets().account(),
+      },
+      fee: {
+        asset: glmr,
+        balance: BalanceBuilder().substrate().system().account(),
+      },
+    }),
+    new AssetConfig({
+      asset: apillon,
       balance: BalanceBuilder().substrate().assets().account(),
       contract: ContractBuilder().Xtokens().transferMultiCurrencies(),
       destination: polkadotAssetHub,